feature: implement open workflows from signature
