+++ nodes/web/signature_api/workflow_services.js
@@ -1,7 +1,6 @@
 import { getAccessToken, signatureApiBaseUrl } from "./main.js";
 
-export const getWorkflowsListForForm = async (element, page = 0, limit = 100) => {
-  const offset = page * limit;
+export const getWorkflowsListForForm = async (element, offset = 0, limit = 100) => {
   const accessToken = getAccessToken();
   const url = `${signatureApiBaseUrl}/api/v1_management/workflow?offset=${offset}&limit=${limit}`;
   const response = await fetch(url, {
@@ -29,6 +28,25 @@ export const getWorkflowsListForForm = async (element, page = 0, limit = 100) =>
 export const getWorkflowById = async (workflowId) => {
   const accessToken = getAccessToken();
   const url = `${signatureApiBaseUrl}/api/v1_management/workflow/${workflowId}`;
+  const response = await fetch(url, {
+    method: "GET",
+    headers: {
+      "Content-Type": "application/json",
+      Authorization: `Bearer ${accessToken}`,
+    },
+  });
+
+  if (!response.ok) {
+    throw new Error(`Failed to get workflow: ${response.status} ${response.statusText}`);
+  }
+
+  const parsedResponse = await response.json();
+  return parsedResponse.result;
+};
+
+export const getWorkflowVersions = async (workflowId, offset = 0, limit = 100) => {
+  const accessToken = getAccessToken();
+  const url = `${signatureApiBaseUrl}/api/v1_management/workflow/${workflowId}/version?includeWorkflow=true&offset=${offset}&limit=${limit}`;
   const response = await fetch(url, {
     method: "GET",
     headers: {
